Make Nvidia GPU preference more granular
Let's have the user decide whether to use the proprietary drivers
or not. The default of "false" is to make sure it all works fine
even on laptops and desktop computers with ancient Nvidia GPUs that
the latest proprietary drivers don't support.

See TODO in pre_tasks/gpu_drivers/all.yml as to what to do next.