ZSH: Fix return code `1` upon launch
When not wrapped around an `if` condition, comparison statements
are assumed as literal commands to be executed.

This shouldn't be that big of a deal with scripts executed directly,
but it's a problem with those `source`d, such as an RC file, as the
last command's return code is reflected to the shell as is.

This is visualized by our customization as a red prefix instead of
pink, which means a command has failed, even though there was no
command run by the user.

Wrapping this around an `if` statement works because the return code
for the statement is only used to decide whether its body should
run or not.

This way returns "0", which means success, if the statement to
check returns failure as there were no real commands run.

If the statement to check returns success, which means the body is
also executed, the last command's exit code is reflected as is.